// src/lib/enum.ts
export enum EExample {
  default = 'default',
  fromFen = 'from_fen',
  lastMoveCrazyHouse = 'last_move_crazy_house',
  checkHighlight = 'check_highlight',
  unselectable = 'unselectable',
  unselectableDraggable = 'unselectable_draggable',
  selectableUndraggable = 'selectable_undraggable',
  initial = 'initial',
  castling = 'castling',
  vsRandom = 'vs_random',
  fullRandom = 'full_random',
  slowAnim = 'slow_anim',
  conflictingHold = 'conflicting_hold',
  move = 'move',
  select = 'select',
  conflictingAnim = 'conflicting_anim',
  withSameRole = 'with_same_role',
  notSameRole = 'not_same_role',
  whileHolding = 'while_holding',
  lastMoveDrop = 'last_move_drop',
  presetUserShapes = 'preset_user_shapes',
  fenAndShapes = 'fenAnd_shapes',
  changingShapesHigh = 'changing_shapes_high',
  changingShapesLow = 'changing_shapes_low',
  brushModifiers = 'brush_modifiers',
  autoShapes = 'auto_shapes',
  visibleFalse = 'visible_false',
  enabledFalse = 'enabled_false',
  customSvg = 'custom_svg',
  in3dDefaults = 'in3d_defaults',
  in3dVsRandom = 'in3d_vs_random',
  in3dFullRandom = 'in3d_full_random',
  autoSwitch = 'auto_switch',
  viewOnlyFullRandom = 'view_only_full_random'
}